Workaround a race condition killing the worker at startup
On an instance I'm taking care of, it happens systematically
that the database is queried before _notify_alive() changes are
visible.

Not being sure of db locking, and transactional behaviour, this is
the simplest cure, but it has the defect of possibly giving rise
to an infinite stream of warnings in cases where the worker should
really die.